More Venator Adjustments
While it was too powerful, I feel like the specific nerfs made took it
too far away from the points that make it what it is- I've re-adjusted
the antimatter back up, but removed the Refit ability (since it was part
of a defunt system), reduced the turbolaser damage a bit, and re-upped
the laser damage a bit as well (most of its armament was laser cannons,
not turbolasers, which was more common in the CW than the inverse, which
is why the later ships have so little laser damage and more Turbo) . It
was also low enough that it was similar to ships half its size, so for
that I'd rather hit its cost efficiency than make ity weaker than the
stuff it *should* be beating. OVerall the effects of the nerfs should be
similar.